Fix a bug with approximativ chunk glue

+++ src/Chunk.hh
@@ -82,7 +82,7 @@ public:
   void generateTexture(const std::string& filename) const;
 
   void generateChunk(const texture_coord_type& coords);
-  int operator()(double x, double y) const;
+  double operator()(double x, double y) const;
 private:
   void createRealCoord(const texture_coord_type& coords);
